+++ lib/startback/caching/entity_cache.rb
@@ -22,12 +22,15 @@ module Startback
     # An EntityCache takes an actual store at construction. The object must meet the
     # specification writtern in Store. The 'cache' ruby gem can be used in practice.
     #
-    # Cache hits, outdated and miss are logged in debug, info, and info severity.
-    # The `cache_hit`, `cache_outdated`, `cache_miss` protected methods MAY be
-    # overriden to change that behavior.
+    # This class supports listeners to track cache hits, misses, outdates and
+    # failures. Listeners can be provided at construction via the options, or by
+    # overriding the `default_listeners` method. The default implementation simply
+    # logs.
+    #
+    # By default, this class raises an error if something goes wrong with the cache.
+    # You can disable this by using the `raise_on_cache_fail` option.
     #
     class EntityCache
-      include Support::Robustness
 
       class << self
 
@@ -40,9 +43,22 @@ module Startback
 
       end # class DSL
 
-      def initialize(store, context = nil)
+      DEFAULT_OPTIONS = {
+
+        # Whether a cache fail raises an exception or not
+        raise_on_cache_fail: true,
+
+        # Default listeners to use, if any. When nil is used, `default_listener`
+        # method is used to create them.
+        listeners: nil,
+
+      }
+
+      def initialize(store, context = nil, options = {})
         @store = store
         @context = context
+        @options = DEFAULT_OPTIONS.merge(options)
+        @options[:listeners] ||= default_listeners
       end
       attr_reader :store, :context
 
@@ -66,6 +82,10 @@ module Startback
         load_entity(pkey).tap{|to_cache|
           store.set(cache_key, to_cache, caching_options)
         }
+      rescue => ex
+        cache_fail(pkey, ex)
+        raise if raise_on_cache_fail?
+        load_entity(pkey)
       end
 
       # Invalidates the cache under a given key.
@@ -77,16 +97,31 @@ module Startback
 
     protected
 
+      def raise_on_cache_fail?
+        @options[:raise_on_cache_fail]
+      end
+
+      def register(listener)
+        @options[:listeners].register(listener)
+        self
+      end
+
+    protected
+
       def cache_hit(pkey, cached)
-        log(:debug, self, "cache_hit", context, op_data: pkey)
+        @options[:listeners].cache_hit(self, context, pkey, cached)
       end
 
       def cache_outdated(pkey, cached)
-        log(:info, self, "cache_outdated", context, op_data: pkey)
+        @options[:listeners].cache_outdated(self, context, pkey, cached)
       end
 
       def cache_miss(pkey)
-        log(:info, self, "cache_miss", context, op_data: pkey)
+        @options[:listeners].cache_miss(self, context, pkey)
+      end
+
+      def cache_fail(pkey, ex)
+        @options[:listeners].cache_fail(self, context, pkey, ex)
       end
 
       def default_caching_options
@@ -150,6 +185,12 @@ module Startback
       # instead.
       def load_raw_data(*args, &bl)
         raise NotImplementedError, "#{self.class.name}#load_entity"
+      end
+
+    protected
+
+      def default_listeners
+        Listeners.new << Logger.new
       end
 
     end # class EntityCache